Adjusted file packaging, implemented more OOP structure, cleansing and reformatting
