+++ src/main/java/storage/Storage.java
@@ -1,11 +1,8 @@
 package storage;
 
-import atom.ui.Atom;
 import exceptions.AtomException;
-import task.Deadlines;
-import task.Events;
-import task.Tasks;
-import task.Todo;
+import task.*;
+import misc.Parser;
 
 import java.io.*;
 import java.util.ArrayList;
@@ -15,17 +12,22 @@ import java.nio.file.Paths;
 import java.util.Scanner;
 
 public class Storage {
-    private static final String DIRECTORY_PATH = "src/main/java/data";
-    private static final String FILE_PATH = DIRECTORY_PATH + File.separator + "atom.txt";
+    private static String directoryPath;
+    private static String filePath;
+
+    public Storage(String directory) {
+        directoryPath = directory;
+        filePath = directory + File.separator + "atom.txt";
+    }
 
     public static void saveTasks(ArrayList<Tasks> taskList) {
         try {
-            Path directory = Paths.get(DIRECTORY_PATH);
+            Path directory = Paths.get(directoryPath);
             if (!Files.exists(directory)) {
                 Files.createDirectories(directory);
             }
 
-            BufferedWriter writer = new BufferedWriter(new FileWriter(FILE_PATH));
+            BufferedWriter writer = new BufferedWriter(new FileWriter(filePath));
 
             for (Tasks task : taskList) {
                 if (task != null) {
@@ -40,46 +42,26 @@ public class Storage {
         }
     }
 
-    public static void loadTasks() {
+    public static ArrayList<Tasks> loadTasks() {
+        ArrayList<Tasks> list = new ArrayList<>();
         try {
-            File file = new File(FILE_PATH);
+            File file = new File(filePath);
             if (!file.exists()) {
-                return;
+                return list;
             }
 
             Scanner scanner = new Scanner(file);
             while (scanner.hasNextLine()) {
                 String line = scanner.nextLine();
-                Tasks task = parseTask(line);
+                Tasks task = Parser.parseTask(line);
                 if (task != null) {
-                    Atom.taskList.add(task);
+                    list.add(task);
                 }
             }
             scanner.close();
         } catch (IOException e) {
             AtomException.storageError("Error loading tasks: " + e.getMessage());
         }
-    }
-
-    private static Tasks parseTask(String line) {
-        String[] parts = line.split(" \\| ");
-        try {
-            String type = parts[0];
-            boolean isDone = parts[1].equals("1");
-            String description = parts[2];
-
-            return switch (type) {
-                case "T" -> new Todo(description, isDone);
-                case "D" -> new Deadlines(description, parts[3], isDone);
-                case "E" -> new Events(description, parts[3], parts[4], isDone);
-                default -> {
-                    AtomException.storageError("Unknown task type: " + type);
-                    yield null;
-                }
-            };
-        } catch (Exception e) {
-            AtomException.storageError("Skipping corrupted line: " + line);
-            return null;
-        }
+        return list;
     }
 }
